fix(pricing): ngưỡng chiết khấu danh mục nhận số lượng lẻ, không còn 500 (GL-07)
category_discounts.min_qty là integer nên Postgres suy tham số số lượng theo kiểu cột: bán 1,5 kg
cho khách có chiết khấu danh mục thì lỗi 500 ở tính giá, tạo đơn và đồng bộ ngoại tuyến.

- Đổi category_discounts.min_qty sang numeric(14,3) trong migration 0062 (USING, giữ dữ liệu);
  đầu vào zod vẫn nhận số nguyên.
- Ép ${quantity}::numeric ở câu so ngưỡng của chiết khấu danh mục và giá theo số lượng.
- PGlite v006 đổi catalog_category_discounts.min_qty sang NUMERIC(14,3), đọc bằng parseQuantity.
- Test: khách có chiết khấu 10%, bán 1,5 kg qua resolve-prices, POST /pos/orders và /sync/push.

+++ packages/shared/src/offline/catalog-store.ts
@@ -150,7 +150,7 @@ export const CATALOG_TABLES: Record<CatalogEntity, TableSpec> = {
       ['customerGroupId', 'customer_group_id', 'uuid'],
       ['discountType', 'discount_type', 'text'],
       ['discountValue', 'discount_value', 'bigint'],
-      ['minQty', 'min_qty', 'integer'],
+      ['minQty', 'min_qty', 'numeric'],
       ['effectiveFrom', 'effective_from', 'text'],
       ['effectiveTo', 'effective_to', 'text'],
       ['isActive', 'is_active', 'boolean'],
